very basic TP/FP/FN strict start pos scoring
We only look for TP, FP, and FN tags.  We ignore data type
mismatches (that is, date vs. named entity vs. etc.). A match
counts if it starts on the same position regardless of where
it ends.  We don't calculate precision, accuracy, etc.  Output
is only to screen.  Ingest formats supported so far are only
i2b2 2016 track-1 and DateAnnot as generated by the basic
UIMA DateTime tutorial.