pattern matching now derived from config files
Separate config files can be provided (with samples included in
this repo) to describe the patterns used to find annotations.

The format of this file follows the generic Python ConfigParser
format for ease of compatability.  Other formats could be supported
later if needed.  A basic sample config has been included for CAS XMI
as generated for the Date Annotation tutorial.  A similar basic
sample config has been provided for i2b2 2016 track 1 formatting.