change arguments because of bug in argparse; added sentences config
A bug in argparse means that argparse has some issues reading in
arguments that are strings starting with a dash.  It tries to parse
them as arguments and not parameters to the last argument.  That
causes a problem in the file suffixing because of how dashes are
used to separate out semantic components.  As a work-around, I
automatically lstrip all arguments to --file-suffix so that
you can pass the argument "-out" by using " -out" instead.

I also changed the way the test suffix is ingested.  It is now a
secondary optional parameter to --file-suffix rather than its
own separate argument + parameter.

A basic sentence scoring config was added to score WebAnno and UIMA
outputs.  There is a new namespace for Clinacuity's features again.

I removed some dead code from scoring_metrics.py.

I now guard against missing test files.  It will treat the file
as completely empty of annotations, thus scoring the full
set of false negatives for the file.