by-type with superclass overprinted scores
When using a superclass for aggregating scores to print, each
entry in the configuration file that matched the superclass was
given its own line, effectively replicating data across lines but
also making performance break-downs not as obvious.

To fix this, I need to unique the list of configuration 'types'
that matched the score-key field before looping over all entries
in the score card to print rows.

I updated the README to include examples along these lines.

I added new tests to verify that the print-string delimiter works,
grouping works by file and by type, and argument passing works.
This last test required explicitly passing the argv[1:] through
the functions so that we could simulate command line arguments
better in the test functions.