gold and test input args now require flag; added write annots to disk
Looking to the future of being able to read in files or pre-extracted
dictionaries, I made it so that the first two arguments are no longer
implicit types.  You now have to supply the --gold-input and
--test-input before the directory path to be explicit and play nicely
later when they are fully optional arguments.

If you provide a --gold-out <dir> and/or a --test-out <dir> arg(s),
then the internal dictionary representation of having processed
a file will be written to disk.  The next stage is to add a flag
that lets you read in said directories and use those directly for
scoring rather than having to reprocess the original marked-up
files.