solved the un-imported pandas bug by using a helper function
I had written a helper function to factor out the score_card/type_counts
data structure creation.  That swallowed with it the direct reference
to pandas (pd) in the score_ref_set function.  I needed to add the
same new_score_card() call to the count_ref_set() function to
clean up that code, as well.