score values can be filtered by a list of regexes
In addition to providing the score-key to pull what field to join
scoring on, you can also filter that list of possible values by
a list of valid values.  The parameter --score-values takes n+
strings that are processed as regexes.  The default value is .*
but you could use "^[DT].*[en]$" to find all score values that
beging with D or T and end with e or n.