basic working offset mapping to neutralize whitespace differences
Many of the different document generations and processing tools
mung whitespace differently.  If the config file contains a path
to the raw content (either a single 'Content XPath' key/value or
'Content XPath' + 'Content Attribute' combination) then we can
extract the original raw document.  The raw document is then split
into a list of chars and we iterate over the list mapping each
non-whitespace character to a new mapping and treating whitespace
chars as null (and thus occupying no space in the new index/mapping).
This offset mapping is passed to the extraction tool so that offsets
recorded in the the XML can be adjusted accordingly.  In essense,
"  Hello , World " and "

Hel
lo , World" are treated as identical because all non-whitespace
characters are in the same order.