new ignore whitespace flag; expanded file dictionary; jupyter stubs
We can explicitly turn on or off the ability to ignore whitespace
in the original raw document.  When on, we'll of course still need
access to the raw source document.  However, even when we have
access to it and _can_ generate a offset_mapping dictionary, we
can use the --heed-whitespace flag to treat the offsets as provided
(with all their whitespace gory details) to be considered valid.
The default performance is to --ignore-whitespace and the flag has
been added for completeness (and in case the default ever changes).

We need to pass the ignore_whitespace argument value on to the
extract_annotations() function, where the real logic runs.

We now want to keep a copy of the (original) raw document in the
output dictionary.  This preservation requires that we make the file
save dictionary more complex.  Instead of just having an annotation
dictionary at the root, we have a dictionary containing, as one element,
the annotations dictionary.  We also have 'raw_content' and 'offset_mapping'.
The raw_content will help in displaying the document for later analysis.
The offset_mapping means that we can sync this document with any
other cognate document with differeing whitespace properties in a
viewer based on shared indices.

The begin_pos and end_pos got renamed from a previous
checked-in-but-not-merged commit.  We'll keep begin_pos and end_pos
as the documents provided offsets.  If we needed to re-map those
offsets based on whitespace issues, then those positions are preserved
in the begin_pos_mapped and end_pos_mapped entries.

Added a jupyter work-horse directory to the ignore list in preparation
for adding lots of jupyter files as mini-tutorials/explanations.