refactor scoring to evaluation positional matches separately
I moved the logic of converting start/end position dictionaries
into TP/FP/TN/FN scores into a separate function.  A large impetus
was to make testing the logic easier for ignoring whitespace vs.
keeping whitespace and how that flag changes how we index
positions.

A major change in analyzing positions is that the begin and end
position are shrunk to the closest non-whitespace character in
case of doing an offset mapping.

We also need to pass the offset_mapping data structures around
to more functions as we're using it to help with crawling
the annotation index under more circumstances.

I also update the tests to include more examples of extracting
document_data and offset_mapping data structures.