added new tests for logic around evaluating positions
Several new tests to check that the annotation dictionary
can be changed to forced an error in position evaluation correctly
based on whether the whitespace flag is on or not.  (Namely,
adjusting the begin_pos doesn't hurt scoring when --ignore-whitespace
but it does when scoring --heed-whitespace.)  I was also able to
reduce passing offset_mapping into the evaluate_positions()
function because we don't need it internally anymore (there).