combined two version of extract_annotations; added raw content check
If we had ignore-whitespace on _AND_ we couldn't find the raw
content for a document, that's a big problem.  I added a new
explicit warning when that happens with advice on how to avoid
the issue.

A previous merge after refactoring introduced two versions of
extract_annotations().  That's a problem.  I combined the best of
both implementations.