updated pre-saved output used in testing
Now that the raw content of a document is more important (and leverage
by the scoring system), I need to update the pre-saved output as
well to reflect changes in the test document.  I alread replaced
the input files used in testing with ZZZZ'd documents (instead of
using "Redacted" as their content string).  Now their sample output
saved-to-disk dictionaries are also updated to the ZZZZ'd variant.