fixed warning msg; flattend ss dict to cover all cases; added fuzzy flag
A cut-and-paste error meant that log messages were warning about
the wrong missing component of a annotation entry.  That is now
fixed to distinguish between missing types, starts, and ends.

Gold and test strict start (annotation) dictionaries are now
flattened so that we score all entries, even when they share a
annotation dictionary key.

Most importantly, the main annotation runner was expanded to include
logic for fully-contained and partial matching.  A new function
for running through gold annotation entries was refactored out of
the original runner.  (More of the logic can be simplified to
clean up the deeply nested if's.)  A fuzzy_flag can be set to
determine the flavor of matching (e.g., exact, fully-contained,
or partial).