optional attributes (prov. in config) added to annotation dict
Files generated by pytest's code coverage tool have been added
to the ignore list.

I refactored extract_patterns from the config file to include
two new helper functions for XPath specific and delimiter
specific patterns.  In the case of XPath patterns, we also
want to have the ability to list arbitrary additional attributes
(comma-separated) that will be passed through to the annotation
dictionary when found.  Some of the common code for both XPath
and delimiter patterns was pulled earlier (before the branch)
to simplify the code base.

Most significantly, the annotation patterns dictionary was
passed into the new helper functions and extract_patterns
but not returned because the variable is edited in place.
Thus, we can treat it as a pass-by-reference rather than
as a pass-by-value.

The new list of optional attributes are passed into the
text extraction file to be pulled from XML files (when
present).

An old verbose flag print message was converted to log.debug.