evaluate_positions no longer returns data; added micro-average term
The string "aggregate" was used in score outputs.  More precisely,
this aggregate score was the micro-average for the corpus.  The output
string has been updated to reflect this (with a few test changes
required).

The function evaluate_positions() no longer returns the data
structure score_card.  We can rely on the pass-by-reference
aspect of passing score_card as an argument to support updates
within the function.