convert --fuzzy-match-flags to list; add depth to score_card
The old command line argument --fuzzy-match now accepts a list
of matching types to score for and is renamed --fuzzy-match-flags.
The default value is still just 'exact'.

The score_card data structure is now one level deeper in order
to accomodate tracking all/any fuzzy matching flags provided.
Instead of score_card directly being a pandas data structure,
it is a dictionary of pandas data structures.  Each key of said
dictionary is the fuzzy flag used to define the matching type.

All relevant tests were strictly updated but not improved upon.
That is, there is at least one print_score_summary() test that
is accidentally passing because the new fuzzy flags are not
being properly looped over in the test or the code.