fixed fuzzy_flags plurality; added tests around command args
I failed to update the fuzzy_flags dictionary entry in args when
I added the ability to provide multiple flags on the command line.
Now the variables is called everywhere by the appropriate plural
version.

To test the above change, I added a little more infrastructure
around the call to main-as-a-script.  There is now an init_args
function that can be called and thus can be tested.  New asserts
added around this function for three logical paths in the
fuction:  default, verbose on, corpus_out defined.