aggregate all metrics by type for macro-averaging
I refactored the value_counts() creation to make multiple references
easier.  We want to both be able to process and sum value_counts.
When the sum of value_counts is greater than zero, at least one
instance of an type is present in reference or system output and
thus it should be included in the macro-average.  Macro-averages
are accumulated until the end, when they are processed.  All
metrics other than TP, FP, TN, FN are divided by the total number
of non-zero occuring entries to get the macro-average.