flag --csv-out supports writing nested tsv to disk
A new output format triggered by --csv-out will generate a nested
csv file that is parallel to the json output file.  This file
structure make direct ingest by many stats programs easier than
the JSON file (which is best for the etude viewer).

Most of the logic of this new output was implemented by replacing
print() lines in print_score_summary() with calles to a new
function output_metrics.  The logic of what gets printed to screen
and what goes to disk is largerly handled within this new helper
function.  Disk writing of the csv is actually further displaced
into a separate function.