added flag for ignoring ref files without a test equiv
Sometimes a test file doesn't exist as a bug and sometimes as a
feature.  Now you can choose whether to treat missing system output
files as a known known (and thus don't score all entries as FN) and
sometimes that reflects a bug in the pipeline (and thus score these
files as missing all annotations).