missed entry when simplifying metric printing for tsv out (cf. 559cb14)
In order to support tsv output, I had to simplify the amount of information
that was always present in a metric list of values.  Thus all references
to metrics[ 1: ] needed to updated to just metrics.  I missed a line,
which caused the first metric to always be skipped over in printing
output files for system output json files.

Tests introduced to prevent changes in test_out and reference_out
folders.  Additional tests to help with csv_out header creation.