filter test and ref configs to be union of each other
Just in case annotation patterns show up in one config but not
the other, we filter the two pattern lists against each other
based on the 'type' key (which is how things get scored down
the line, anyway).  The current approach is to use a strict
union requirement.  The pattern type must be in both test and
ref for it to be in either.  Warnings are generated for all dropped
annotations (to help you in case it was an accident).  Errors are
generated if either annotations list ends up being empty after
the filtering.