updated --delim options; explicit print to screen for all numbers
I added a new variant for -d to be --delim.  Also there is a
--delim-prefix option to prefix a different string at the start
of ever line going to stdout.  (This is helpful for formatting
tables in markdown and emacs org-mode.

I replaced --count-types with a full set of --print-counts, --no-counts,
--print-confusion-matrix, --no-confusion-matrix, --print-metrics,
and --no-metrics.  Counts and confusion matrix are off by default
and metrics are on my default.  The full set of commands is available
to make command line calls clear (despite any changes to defaults
in the future).

__main__ in etudy.py was simplified to allow for confusion_matrix
and score_summary printing to be called from the same score_ref
function.  Because the --count-types was removed as an option,
the initial logic of what to run was also simplified.