refactor comparison runner into a three-pass performance filter
In order to simplify each pass/fail step of logic, I refactored
the comparison runner into a three-pass performance filter.
First, we look for clear exact matches for start/end and score
them as pass/fail based on their type matching or not.  Then,
if we're doing other fuzzy flags and the reference annotation
didn't already have an exact match, we try to match things
according to fully-contained logic.  Then, if we're doing partial
matching and the ref didn't match fully-contained, we look for
a partial match.  This still doesn't suppor the full range of
overlapping and nested and many-to-many matching but it will
make creating that logic easier (I think).

Some of the tests had to be updated to accomodate different
return orders of the matches but the scores stay the same.