moved argument lstrip-ping to a single point
Arguments that can take multiple strings and could possibly start
with a dash (like --file-suffix) have to be modified so that
the argparser can handle them and then re-modified so that they work.
To pass by argparser, it's required to prefix any arguments that
would start with a - with a space ("-04.xml" -> " -04.xml"). Then
those entries are modified by .lstrip() of the extra spaces to
be what is intended.  I moved the lstrip() to the init_args stage
so that any in-code usage can assume the arguments are semantically
as intended and don't require any extra pre-processing.