relax required status for ref and test input folders with tests
A new check is added to require reference_input and test_input arguments
programatically rather than by setting the required=True flag in the
arg parsers so that we can have more complex notions of required.
Namely, both arguments are required only when scoring documents
and not when doing a simple print-counts.  I'll need to tighten
the flags up more or make it logically impossible to score files
if both are provided later.