print counts and metrics as both possible; no file-specific output for counts
Originally, print_counts and print_metrics were two sides of a single
if/elif.  We want both to be possible in the same run so convert
the if/eli to a if...if.

Don't print any file-specific output for print_counts.  It's all
all the corpus level and printed to screen.