add explicit tests for print_counts output to csv and screen
I was able to simplify some of the test using mock patch for
reading in command line args so we only imported once at the
top of the file.

Sample out for csv for print_counts added and explicit tests to
compare this output to print_counts.  stdout printing of counts
also tested with inline examples.