casting start/end positions as int when possible
Originally, we had treated all start/end annotation positions as
strings.  Some where int (as document offset positions) and others
were strings representing other positions types (specifically, SOF
for start of file and EOF for end of file, although others could
be used later).  Matching in the partial alignment runner had a
leak such that the greater than and less than comparisons could
break if the wrong strings happened to be compared.  For instance,
the string order of '12' is greater than the string order of '116',
would would lead to a false overlap.

Now, we try to cast the position as an int, if that fails, keep
it as as string.  This doesn't seem to introduce any other comparison
problems.

Lots of tests had to be updated as well because the gold standard
was stored as a '12' rather than as 12, etc.