pivot_attr added to config
An xml pattern defined in the config can include pivot_attr instead
of start_pos and end_pos attributes.  This is intended for
document level (spanless) annotations.  It will use the XPath
to match an XML element and then extract the value from the provided
attribute.  That attribute value is treated as the value of the
annotation.  For example, use XPath to target the Author tag and
the pivot attribute to target the value containing the actual
name.