skip nulls more precisely; --empty-value for output
Original logic didn't include in macro-averages any
files or types that had zero annotations across the board. More
precise exclusion from averages requires that we only ignore those
with a divide by zero or non-computable value.  That is, precision
is null when TP and FP are zero.  FN and TN don't matter. This update
means we distinguish between zero and null/none in the scoring script
printer.  null/nones are always rendered as None (or null in json
output) and have a roll up effect (that is, f1 is null if precision
or recall is null).

I added a flag to let you set what empty or null or non-computable
values will show up as in the output.  For now, this flag doesn't
do anything but it serves as a hook for later usage.