simple --pretty-print support for metrics output
When printing score metrics, all numerical columns are forced to
be padded up to 9 chars.  Things that look like ints are treated
as ints (even 1.0 in the F1 column, for example). Things that look
like floats are rounded to four digits. A horizontal rule under
the header is added for readability.  --delim " " (space) is a
better default or even --delim " | ".

Fancier algorithms and/or configurable algorithms can be added later.