converted to RST
Converted the README file from markdown to restructuredtext as part
of a migration to Sphinx for easier documentation.

Added some simple docstrings to functions to help frame out the
basics of the Sphinc doc tree.

When writing JSON structures to disk, sort the indices (when
possible) to make regression testing easier.