basic working model to print attribute/pivot scores
Attributes on annotations are treated as pivots (much like the
spanless/document level annotations).  These pivots are now
printed for regular annotations and not just
these special types. Many values are still hard-coded so this
commit is a work-in-progress stage to make it easy to back-track.

We also get some speed-ups in file scoring and type scoring
because of a slight shift in when the algorithm checks for
--by-file and --by-type flags.