score attributes flag
A new --score-attributes flag allows you to optional score any
attributes that overlap between the reference and system out
pattern files. Not using the flag means all attributes are tracked
but not scores.  Providing just the flag alone means that all
attributes in the intersection of the reference and system out
patterns are scored.  Otherwise, a string listing attributes delimited
by commas servers as a filter for the intersection of the possible
attributes.