f-beta value flag; F2, F0.5 metrics
New simplified/syntactic sugar options for metrics F2 and F0.5 for
beta values of 2 and 0.5, respectively provided to complement
the normal F1 metric.  Any and all beta values can also be provided
via the --f-beta-values <value list> flag, such as:
-m F --f-beta-values 1 2 0.5 0.34 4

We do some pre-processing in etude.py to align the combination of
arguments to the metrics list and the beta values list.

Lots of tests added.