iterate over dictionary keys as per py2 and py3; replace in.keys()
Iterating over dictionary keys needed to be made comptible with
both py2 and py3.  Also the default method for looking to see
if a key was in a dictionary (when used in conditional vs.
loops, as above) could be simplified.

When iterating over the keys of a dictionary, we don't want to
use the .keys() options to stay compatible between py2 and py3