unit test error with findall( namespaces); more dict.keys clean-up
add_missing_fields is an exception to the .keys removal because
the datastructre in question is actually a pandas datastructure
rather than a python dictionary.

I missed a few dictionary.keys entries.

A unit test error came to light because a list.items and a dict.keys
were treated similarly for namespaces.