fixed specificity calc, empty_value; added parity and custom print
The original definition of specificty was off because it used FN
rather than FP in the calculation.

The default empty_value need to get standardized to make fancy
printing easier and more robust.

Updates to the gitignore file are tied to changes in how I run
pytest as a standard to make it easier to inspect different
version of python.

The new config file attribute Parity is used currently as a
cleaner signal that we're talking about a document-level
annotation.  It will also be used to distinguish between
those annotations that need to be unique to a file, required
for a file, or could occur any number of times in a file.

The first sample custom print (that is, a print score
that doesn't rely on the standard configurable output) is
introduced to mock up the 2018 n2c2 track 1 output. The default
scoring output can still be used but not all the values are
easy to generated in that format.