simple working extract_samples for py3 compliance
py2 runs into encoding issues to be ironed out later (maybe).

There are still some newline chars that I need to get rid of
even for this py3 version.

You need to point the script to a score card and a system
output folder.  Optionally, you can specify which metrics
to filter for and how large a margin you want on the left and
right side of the annotation.