updated all README examples; fixed CAS_XMI.conf to new reqs; patched config parser error
The README now shows all the outputs to match exactly as they run in
the current output.  Little bug patches and extensions here and there
had caused the actual output to drift from the README. Likewise,
the CAS_XMI.conf example config ened to be updated to match current
parsing requirements.

In doing so, I found a new edge case for using the Long Name/Section
as the score key and refactored some of the logic around that to make
sure we don't accidentally hit an undefined variable.